[r505] [branches/dev] Work-in-progress refactoring of DisplayFrame in Chatterbox. DisplayFrame has been made much more object-oriented. It is not yet functional, though, as it relies on new concepts in the model that are not yet implemented.
DisplayFrame has been replaced by ChatterboxWidget, a Polish Container that embodies the Frame. A ChatterboxWidget is a view tied to a single question in the model. It contains the basic scaffolding to toggle the widget between an 'expanded' and 'collapsed' mode. What actually happens in these modes is determined by strategy objects passed in via the ChatterboxWidget constructor.

The contents of the ChatterboxWidget are determined by strategy objects that implement IWidgetStyle (for collapsed, read-only mode), and IWidgetStyleEditable (for expanded, editable/interactive mode). These strategy objects know how to construct the widget (add items to the ChatterboxWidget container), keep the widget up to date in response to changes in the question, and (optionally) get data back out of the widget.

WidgetStyle implementations are provided for most (but not all) of the common styles currently available in Chatterbox.

A new concept of a 'question data object' has been created that these widgets rely on. This concept encapsulates the answer to a question, in which the object contains an internal representation of the answer data, as well as methods to render the answer data into human-readable form (for the GUI), and back-end forms like XML. This gets this logic out of the DisplayFrame (and other scattered places) where it did not belong. This idea probably overlaps with Daniel's implementation and will have to be merged.