Delete excessive repeats driven by jr:count attr
The problem description:
Repeat groups that are controlled by the jr:count attribute which is
a path to a node (a number question) that holds the actual value
do not respect situation where the value is *decreased*.

Example:
User starts a form and puts 6 as the answer for the question that
controls the jr:count of the repeat group.
As (s)he is going forward, the repeat groups are properly created
by FormEntryModel.
The user goes back to the number question and changes it to 3.
As (s)he is going forward, the new value is ignored
by FormEntryModel and the number of repeat groups is still 6.

Solution:
The most of the magic takes place in
FormEntryModel#createModelIfNecessary(FormIndex) method which takes care
of creating the missing repeat groups by comparing the count value with
the currently being processed group instance index. If the index is
lower than the count value and the group does not exist, the method
creates it.
Now the method also checks if the instance index is *higher* than the
count value. If so, it deletes the currently processed group and all
the following (if any) in a loop.