Skip irrelevant indices after deleted repeat group
If there is at least one repeat group to be deleted
during updating a model, user will be taken to the first form index
after the deleted repeat group(s). This happens after FormEntryController
already skipped non-relevant indices so the skip must be repeated if
the the first index after the deleted repeat group(s) is non-relevant
(hidden).