Fix round(num, power) for some locales.
This function does not work on locales that use
comma as decimal separator.
XPathFuncExpr#eval throws NumberFormatException.

It happens because NumberFormat class is used to
make the rounding. This class uses the current
locale's settings so US (or any other that use dot
as decimal mark) locale must be forced.