Renamed SMSSerializingVisitor to CompactSerializingVisitor
This was done because this current version of the visitor isn't compatible with forms that were utilizing the previous version of SMSSerializingVisitor before these changes were made. Also the choice to name the class CompactSerializingVisitor was made since the spec and doesn't stipulate anything that makes this visitor SMS specific.