Allow references with ../ in them to be anchored
The parent method can't anchor references with parent steps but anchor can. This fix allows us to run current_in_itemset_nodeset_should_refer_to_node because the form can now be parsed. We now get an error because there's an attempt to bind the repeat to /people which doesn't exist and is also not possible because it would be the root instance. This is because current() in the itemset nodeset reference is being expanded as /data instead of /data/selected_person.