Migrate assertions to something that's easier to reason about
Using epoch values was working but it's harder to understand by humans than actual time strings.

Also, by using OffsetDateTime values and parsing strings with java.time we are ensuring as well that our test values are correctly parsed from string to date times because java.time is a battle tested Java API.

Now the only thing that could be wrong in the tests are the input and output time strings (both in human readable format, easy to check) and the offset definitions.