Migrate Scenario inner workings to use the form entry controller
The Scenario tool has two different APIs: one that lets test authors traverse the form as the user would do with an ODK client, and another to surgically write answers in specific xpaths.

The first API uses the form entry controller, and the second API directly mutates the main instance.

This commit tries to unify both APIs to use the form entry controller, which is safer and also provides examples about how a form can be used to new developers.

The previous implementation has also a separate algorithm to create TreeReference instances from xpath Strings, which didn't behave exactly as our XPath parser. Now we use the parser to be closer to what happens in the implementation.