Improve jr:choice-name dynamic itemset test
This test was the only one that failed when I changed the Scenario class to always work off of a deserialized form definition. It turned out to have a misunderstanding about when choice lists are evaluated. The test now better matches client behavior and is consistent pre and post serialization.