Identify triggerables that affect all repeat instances
For performance reasons, this needs to only be done if we know the changed reference is a repeat. It's not possible to quickly identify whether a TreeReference represents a repeat/repeat instance so we need some information from createRepeatInstance. We could pass a boolean representing whether the changed node is a repeat but this seems clearer to me.