Add test with sum and fix contextualize
The new test highlighted that the condition to ensure generic reference steps were respected in contextualization was incorrect. Parenting steps are evaluated against the parent ref, not the newly-created ref. The tests didn't previously highlight this because all cases were when the new ref was one step shorter than the context ref. The new expression is `i < contextRef.size() - refLevel`. This makes sense semantically: we want to consider multiplicity from the context ref only for steps that we haven't parented past. refLevel represents the number of parenting steps in the reference to contextualize. `contextRef.size() - refLevel` gets us to the absolute part of the contextRef that is exactly duplicated in the new, contextualized reference.