Serialize and deserialize namespace prefix
An alternative would be to serialize the namespace map. Currently this is only available at parse time so it would be a bigger change. We generally don't expect many namespaced instance elements so this is a pragmatic approach for now. If we find namespace usage takes off, we can reconsider.