Cache predicate evals for single triggerable evaluation runs
Only do this for eq predicates and avoid expressions with
functions. Allowing this for any kind of expression causes tests
to fail.

The caching is also limited to triggerable evalutions so that it
doesn't kick in other cases like nodeset calculations for selects.
Caching during these cases also causes problems (probably because
this cache isn't being cleared in those cases).